Merge ScopedVisitor with ScopeResolverVisitor classes.
Now that ResolverVisitor no longer deals with scopes,
ScopeResolverVisitor is the only class that inherits from
ScopedVisitor, so we can combine the two.  This makes the code a lot
easier to read and reason about, since we don't have to bounce between
the two classes with super calls all the time.

Change-Id: I308edd1a7181001724b89ffb04499b9a3df25cb6
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/209852
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>