Add macroGeneratedContent, set nameOffset(s) for macro-generated elements.
The idea is that for macro-generated elements navigation point at
a temporary file that contains this generated content, and so name offsets
should be in this file. Elements that were produced from "source" code,
should be in the "source" file, with "source" name offsets.

Change-Id: Iefa105299444fbadfd3d52048c4dfe379b482449
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/209480
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Samuel Rawlins <srawlins@google.com>