Support navigation to macro-generated elements.
We use previously added `CompilationUnitElementImpl.macroGeneratedContent`
to access the current combined content.

In order to make this work with current IntelliJ plugin, without
adding new protocols, we store the generated code into a new file,
and locations for macro-generated elements are in this file.

Locations for elements that are not macro-generated are still
in the source file.

We use `ElementLocationProvider` to separate writing to the file
system, and do this only in DAS.

Change-Id: Ib156c67b0bf1f3474eb06584e76aa532abd597d1
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/210300
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>