[vm] Fix a false-positive keeping switchable calls in the monomorphic state.
When checking if the guarded receiver CID has changed, we must find the original guarded CID by looking in the original ICData entries array. The first entry in the current ICData entries array might not correspond to the original entry because we sometimes re-order entries. For example, insert the Smi case at the beginning.

TEST=ci
Bug: https://github.com/dart-lang/sdk/issues/46948
Bug: https://github.com/dart-lang/sdk/issues/46980
Change-Id: Id0904de3a68fbb77842f552329a14c3127ee9fc8
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/211082
Commit-Queue: Ryan Macnak <rmacnak@google.com>
Reviewed-by: Alexander Aprelev <aam@google.com>