analyzer: constructor tearoffs can be const.
Fixes https://github.com/dart-lang/sdk/issues/46899

Also required fixing the dangling type parameters on
ConstructorElements, FunctionTypes and the returnType (an
InterfaceType) of the FunctionTypes. In the case of the
ConstructorElements, we just refer to the declaration.

Fixes https://github.com/dart-lang/sdk/issues/46981

To remove dangling type parameters, we set
`constructorName.type.type` to null for generic ConstructorReferences.
The idea is that this code is generic and the TypeName is not a proper
type. This requires a few changes to other code that expects
TypeName.type to be non-null.

Change-Id: Ie14344c5179922318ca746cbbdab13cf67b7432a
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/211304
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Samuel Rawlins <srawlins@google.com>