Separate index/search kind for ConstructorReference.
When we just search, it is enough to know that there is just any
reference kind, however wheb I attempted to implemented rename
refactoring, I realized that we need to know if the reference is
from a ConstructorReference, so `.new` is required.

I think this change will let us do this, however I still have not
tried to actually wire it, so might still be wrong.

Change-Id: Ie97e23d391ea89d7f3695663368835e747cbeee5
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/212546
Reviewed-by: Samuel Rawlins <srawlins@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>