Detect changes to and always use the bazel-bin symlink
Previously if we found a file was generated using one Bazel
configuration, then we'd continue using its path even if a new build
with a different configuration was more recent. In other words, we'd use
a stale generated file.

This commit gives `bazel_watcher.dart` the ability to detect changes to
symlinks which means that we'll do a refresh if `blaze-bin` changes
(it's already being watch in `context_manager.dart`). But since
`binPaths` previously always looked at all the possible "bin paths",
even after a refresh we could end up with using the wrong one. So the
second change of this commit is to look only at the `blaze-bin` symlink
in `findFile` (the symlink is updated by Blaze) to make sure that we
pick only the most recent path. This doesn't change `binPaths` itself
as it's also used for other purposes.

Bug: b/193636339
Change-Id: I641a0c9a5694121435360ca46b275d0b920d2911
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/209701
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Michal Terepeta <michalt@google.com>