Fix error message text for INVALID_MODIFIER_ON_SETTER.
This error message was previously reported by the analyzer's parser,
which presumably supplied an argument describing the modifier.  It is
now reported by the shared parser, which doesn't supply any additional
information.  As a result, the error message was showing up for the
user as "The modifier '{0}' can't be applied to the body of a setter."
(with a literal `{0}` visible to the user).

Changed the error message text to match what the CFE reports.  There's
no real need for the message to describe the modifier since the error
message location is the modifier itself.

Change-Id: I538d49c1b693fe6cfb190f584bb5c9ab5cd3cf39
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/213732
Reviewed-by: Samuel Rawlins <srawlins@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>