Fix reporting of HintCode.INVALID_VISIBLE_FOR_OVERRIDING_ANNOTATION
The hint code text is:

    The declaration '{0}' is annotated with 'visibleForOverriding'. As
    '{0}' is not an interface member that could be overriden, the
    annotation is meaningless.

So when instantiating this error we should only provide a single
argument.  Previously we were providing a second argument that was the
name of the annotation; this is unnecessary because the error message
already mentions the annotation name ('visibleForOverriding').

Change-Id: I022af6914668ad6232005237619b3ab9a5d7a589
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/213825
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>